feat: integrate forecast AQI, add mandir-marg zone, and UI polish
- Frontend (Dashboard.jsx): Improved state management by explicitly clearing forecast and attribution data on zone switches. Updated current AQI to use forecasted values if available. Increased enforcement priorities fetch limit to 50 while rendering the top 10.

- Backend/ML: Added mandir-marg to VALID_ZONES and zones_metadata.csv. Updated enforcement ranking to preferentially use forecasted AQI. Fixed a type issue in forecasting/model.py by casting improvementPercent to a float.

+++ ml-service/forecasting/model.py
@@ -120,7 +120,7 @@ def get_forecast(ward_id: str) -> dict[str, Any]:
             "modelRMSE": float(model_rmse),
             "persistenceRMSE": float(persistence_rmse),
             "modelName": f"SARIMA-Seasonal-{current_season.capitalize()}",
-            "improvementPercent": round((1-model_rmse/persistence_rmse)*100, 1) if persistence_rmse > 0 else 0.0,
+            "improvementPercent": float(round((1-model_rmse/persistence_rmse)*100, 1)) if persistence_rmse > 0 else 0.0,
         },
         "dataSource": "pre-trained-pkl-models",
     }
